requirements.txt
absl-py==1.4.0
alembic==1.10.3
anyio==3.6.2
astunparse==1.6.3
asyncpg==0.27.0
auth==0.5.3
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
dnspython==2.3.0
email-validator==1.3.1
eventlet==0.33.3
falcon==3.1.1
fastapi==0.95.1
fastapi-users==10.4.1
fastapi-users-db-sqlalchemy==5.0.0
flatbuffers==23.3.3
gast==0.4.0
geographiclib==2.0
geopy==2.3.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
lxml==4.9.2
makefun==1.15.1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
mongoengine==0.27.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.8.10
packaging==23.1
passlib==1.7.4
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.9
starlette==0.26.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
Werkzeug==2.2.3
wrapt==1.14.1
